Move enums to shared available data instead of specific receipt
