+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Domain/AvailableChargeReceiptData/AvailableChargeReceiptDataReasonCode.cs
@@ -14,6 +14,7 @@
 
 using System;
 using GreenEnergyHub.Charges.Domain.AvailableChargeLinkReceiptData;
+using GreenEnergyHub.Charges.Domain.AvailableData;
 
 namespace GreenEnergyHub.Charges.Domain.AvailableChargeReceiptData
 {
