+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Domain/AvailableData/ReasonCode.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace GreenEnergyHub.Charges.Domain.AvailableChargeLinkReceiptData
+namespace GreenEnergyHub.Charges.Domain.AvailableData
 {
     public enum ReasonCode
     {
