+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Infrastructure/ChargeLinkReceiptBundle/Cim/ChargeLinkReceiptCimSerializer.cs
@@ -16,6 +16,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Xml.Linq;
 using GreenEnergyHub.Charges.Domain.AvailableChargeLinkReceiptData;
+using GreenEnergyHub.Charges.Domain.AvailableData;
 using GreenEnergyHub.Charges.Domain.MarketParticipants;
 using GreenEnergyHub.Charges.Infrastructure.Cim;
 using GreenEnergyHub.Charges.Infrastructure.Configuration;
