+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Infrastructure/Context/EntityConfigurations/AvailableChargeLinkReceiptDataConfiguration.cs
@@ -14,6 +14,7 @@
 
 using System;
 using GreenEnergyHub.Charges.Domain.AvailableChargeLinkReceiptData;
+using GreenEnergyHub.Charges.Domain.AvailableData;
 using GreenEnergyHub.Charges.Domain.MarketParticipants;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
